ダッシュボード機能の`TopView`コンポーネントを`TodaysOverview`に置き換え、不要な`LinksOverview`および`SwipeOverview`コンポーネントを削除しました。これにより、コードの整理と可読性が向上しました。
